extract some base images to improve ci build

+++ docker/builder/build.sh
@@ -1,22 +1,29 @@
 #!/bin/bash
 
+SCRIPTPATH="$( cd "$(dirname "$0")" ; pwd -P )"
+SRCPATH=${1:-$(cd $SCRIPTPATH/../..; pwd -P)}
+NPROC=${NPROC:-$(nproc || grep -c ^processor /proc/cpuinfo)}
+
 set -xe
 
-mkdir -p /server/build_docker
-cd /server/contrib/kvproto
-rm -rf cpp/kvproto
-./generate_cpp.sh
-cd /server/build_docker
-cmake /server -DENABLE_EMBEDDED_COMPILER=1 -DENABLE_TESTS=1 -DCMAKE_BUILD_TYPE=Debug
-make -j $(nproc || grep -c ^processor /proc/cpuinfo)
+if [ -d "$SRCPATH/contrib/kvproto" ]; then
+  cd "$SRCPATH/contrib/kvproto"
+  rm -rf cpp/kvproto
+  ./generate_cpp.sh
+  cd -
+fi
+
+build_dir="$SRCPATH/build_docker"
+mkdir -p $build_dir && cd $build_dir
+cmake "$SRCPATH" -DENABLE_EMBEDDED_COMPILER=1 -DENABLE_TESTS=1 -DCMAKE_BUILD_TYPE=Debug
+make -j $NPROC
 #ctest -V -j $(nproc || grep -c ^processor /proc/cpuinfo)
 
-install_dir="/server/docker/builder/tics"
-mkdir -p "$install_dir"
-rm -rf $install_dir/*
-cp -f "/server/build_docker/dbms/src/Server/theflash" "$install_dir"
+install_dir="$SRCPATH/docker/builder/tics"
+if [ -d "$install_dir" ]; then mkdir -p "$install_dir"; else rm -rf "$install_dir"/*; fi
+cp -f "$build_dir/dbms/src/Server/theflash" "$install_dir"
 
-ldd "/server/build_docker/dbms/src/Server/theflash" | grep '/' | grep '=>' | \
+ldd "$build_dir/dbms/src/Server/theflash" | grep '/' | grep '=>' | \
   awk -F '=>' '{print $2}' | awk '{print $1}' | while read lib; do
   cp -f "$lib" "$install_dir"
 done
